// src/app/models/persona/Persona.ts
export class Persona {

    id: number;
    nombre: string;
    apellido: string;
    fechaNacimiento: Date;
    edad: number;
    mensajeCumpleano: string;

/*     constructor(id: number, nombre: string, apellido: string, fechaNacimiento: Date, mensaje: string, edad?: number) {

        this.id = id;
        this.nombre = nombre;
        this.apellido = apellido;
        this.fechaNacimiento = fechaNacimiento;
        this.mensaje = mensaje;
        this.edad = edad;
    } */
}